refactor: modules directory
